airflow-pipeline-with-spark/pyproject.toml
[project]
name = "airflow-pipeline-with-spark"
version = "0.1.0"
description = "Airflow pipeline with PySpark for data processing"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.3",
    "apache-airflow-providers-apache-spark>=5.3.4",
    "pandas>=2.3.3",
    "pyspark>=4.0.1",
    "pytest>=9.0.1",
    "requests>=2.31.0",
]
